requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch
transformers==4.37.2
Pillow
python-multipart
numpy
requests
python-dotenv